# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: WebChat Public API
  version: 0.4.0
  description: >
    Публичный API веб-чата. Единственный источник истины публичного API;
    TypeScript-типы фронтенда генерируются из этого файла автоматически.

    Фича 002 (auth): endpoints №1–10 — регистрация с email-подтверждением,
    вход по паролю, refresh-ротация, logout, восстановление пароля, /users/me.
    Публичны ровно endpoints №1–6, №8–9 (8 путей); №7 logout и №10 /users/me
    требуют Bearer. Все ответы об ошибках — RFC 9457 application/problem+json
    (схема Problem). Ссылки в письмах указывают на маршруты SPA, а не на API.

    Фича 003 (sso): endpoints SSO §1–7 — вход через внешних
    identity-провайдеров (OIDC/OAuth2): перечень провайдеров,
    authorize/callback/token, привязка и отвязка идентичностей. Дополнительно
    публичны 4 маршрута (providers, authorize, callback, token);
    link/authorize, GET /users/me/identities и DELETE идентичности
    требуют Bearer. Endpoint'ы 002 не изменились (аддитивно, minor-бамп).

    Фича 004 (messaging): endpoints №11–24 — личные текстовые диалоги
    (ensure/чтение/удаление, история с курсорной пагинацией, отправка с
    клиентской дедупликацией, read-отметки), реалтайм-канал SSE
    №18 /users/me/events, контакты, точный поиск пользователей и
    блокировки (№19–24). Все новые пути требуют Bearer: публичный
    перечень 002–003 не расширен. Фиксированные константы: максимум текста
...
    только additive (0.3.0 → 0.4.0, без BREAKING.md).
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: Список контактов пользователя (№20, Bearer)
      description: >
        Контакты вызывающего с серверной алфавитной сортировкой по
        выбранному полю без учёта регистра (FR-015): sort=login — по
        lower(username) (по умолчанию), sort=email — по lower(email).
        Пустой список — корректный ответ.
      security:
        - bearerAuth: []
      parameters:
        - name: sort
          in: query
          required: false
          description: 'Поле серверной алфавитной сортировки без учёта регистра; по умолчанию login; иное значение → 400 invalid_sort'
          schema:
            type: string
            enum: [login, email]
            default: login
      responses:
        '200':
          description: Контакты, отсортированные алфавитно по выбранному полю без учёта регистра
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactsResponse'
        '400':
          description: 'sort не login и не email (errors: {sort: [invalid_sort]})'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '401':
          description: Не аутентифицирован (нет токена / истёк / отозван / недействителен — единообразно)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    post:
      operationId: addContact
      tags: [contacts]
      summary: Добавить пользователя в контакты (№21, Bearer)
      description: >
        Односторонняя запись в личный список контактов вызывающего
        (FR-016). 201 — контакт создан; 200 — уже добавлен, дубль не
        создаётся (edge). Добавление самого себя → 422 self_forbidden.
        Контакты не зависят от чатов и блокировок: удаление контакта (№22)
        не затрагивает единый диалог пары (FR-017); блокировка не
        затрагивает контакты (FR-020) — заблокированный может добавлять
        блокирующего в свои контакты.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [userId]
              properties:
                userId:
                  type: string
                  format: uuid
                  description: Добавляемый пользователь; добавление самого себя → 422 self_forbidden
      responses:
        '200':
          description: Контакт уже добавлен — возвращён существующий ContactView (дубль не создаётся)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactView'
        '201':
          description: Контакт создан — ContactView
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactView'
        '401':
          description: Не аутентифицирован (нет токена / истёк / отозван / недействителен — единообразно)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: 'Пользователь не найден (errors: {user: [user_not_found]})'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: 'Добавление самого себя в контакты запрещено (errors: {userId: [self_forbidden]})'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/contacts/{userId}:
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Удалить контакт (№22, Bearer)
      description: >
        Удаляет пользователя из списка контактов вызывающего. Идемпотентно:
        несуществующий контакт — тоже 204. Удаление контакта — независимая
        операция: единый диалог пары и его история не затрагиваются,
        переписка остаётся доступной и продолжается (FR-017).
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          description: Удаляемый из контактов пользователь
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Контакт удалён (идемпотентно; чат и история пары не затронуты)
        '401':
          description: Не аутентифицирован (нет токена / истёк / отозван / недействителен — единообразно)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Authorization: Bearer <accessToken>'
  schemas:
...
    PublicUser:
      type: object
      additionalProperties: false
      required: [id, username, email, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
          minLength: 3
          maxLength: 32
          pattern: '^[a-zA-Z0-9]([a-zA-Z0-9_.-]{1,30}[a-zA-Z0-9])$'
        email:
          type: string
          format: email
          maxLength: 254
        status:
          type: string
          enum: [pending_email_confirmation, awaiting_password, active]
        createdAt:
          type: string
          format: date-time
    Problem:
      type: object
      description: 'RFC 9457: единая модель ошибок; расширения допускаются'
      required: [title, status]
      properties:
        type:
          type: string
          description: URI-reference на тип ошибки (опционально)
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        errors:
          type: object
          description: Коды ошибок валидации по полям (в 400-ответах)
          additionalProperties:
            type: array
            items:
              type: string
              description: Код ошибки, например invalid_format
...
    ContactView:
      type: object
      additionalProperties: false
      required: [user, createdAt]
      properties:
        user:
          $ref: '#/components/schemas/PublicUser'
          description: Пользователь из списка контактов вызывающего
        createdAt:
          type: string
          format: date-time
          description: Когда пользователь добавлен в контакты вызывающего
    ContactsResponse:
      type: object
      additionalProperties: false
      required: [contacts]
      properties:
        contacts:
          type: array
          description: 'Контакты вызывающего: серверная алфавитная сортировка по выбранному полю без учёта регистра — lower(username) при sort=login, lower(email) при sort=email (FR-015); пустой список валиден'
          items:
            $ref: '#/components/schemas/ContactView'
...